Allow passing data to transmogrify
Instead of only allowing URLs and file paths as an argument to transmogrify this also allows passing raw data as a javascript object.

The use case for this is being able to have transmogrify run multiple times based on only one given resource. Using this I can fetch whatever is needed, split it into parts and call transmogrify on those parts individually.